+++ nuclia/sdk/upload.py
@@ -18,22 +18,12 @@ from nuclia.decorators import kb
 from nuclia.lib.conversations import Conversation
 from nuclia.sdk.auth import NucliaAuth
 from nuclia.sdk.logger import logger
-
-RESOURCE_ATTRIBUTES = [
-    "icon",
-    "origin",
-    "extra",
-    "conversations",
-    "texts",
-    "usermetadata",
-    "fieldmetadata",
-    "title",
-]
+from nuclia.sdk.resource import RESOURCE_ATTRIBUTES, NucliaResource
 
 
 class NucliaUpload:
     """
-    Create or update resource in a Nuclia KnowledgeBox.
+    Create or update resource content in a Nuclia KnowledgeBox.
 
     All commands accept the following parameters:
     - `rid`: Resource ID. If not provided, a new resource will be created.
@@ -187,6 +177,32 @@ class NucliaUpload:
             )
 
     @kb
+    def link(
+        self,
+        *,
+        uri: str,
+        **kwargs,
+    ):
+        """Upload an URL to a Nuclia KnowledgeBox."""
+        field = kwargs.get("field") or uuid4().hex
+        links = {
+            field: {
+                "uri": uri,
+            }
+        }
+        kwargs["icon"] = "application/stf-link"
+        rid, is_new_resource = self._get_or_create_resource(
+            links=links,
+            **kwargs,
+        )
+        if not is_new_resource:
+            self._update_resource(
+                rid=rid,
+                links=links,
+                **kwargs,
+            )
+
+    @kb
     def remote(
         self,
         *,
@@ -228,7 +244,7 @@ class NucliaUpload:
     def _get_or_create_resource(*args, **kwargs) -> Tuple[str, bool]:
         rid = kwargs.get("rid")
         if rid:
-            return rid
+            return (rid, False)
         ndb = kwargs["ndb"]
         slug = kwargs.get("slug")
         need_to_create_resource = slug is None
@@ -258,12 +274,4 @@ class NucliaUpload:
         return (rid, need_to_create_resource)
 
     def _update_resource(self, rid: str, **kwargs):
-        ndb = kwargs["ndb"]
-        kw = {
-            "kbid": ndb.kbid,
-            "rid": rid,
-        }
-        for param in RESOURCE_ATTRIBUTES:
-            if param in kwargs:
-                kw[param] = kwargs.get(param)
-        ndb.ndb.update_resource(**kw)
+        return NucliaResource().update(rid=rid, **kwargs)